requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asn1crypto==1.5.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloud-sql-python-connector==1.18.1
cryptography==44.0.2
dnspython==2.7.0
frozenlist==1.6.0
GDAL==3.10.3
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1rc1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
greenlet==3.2.1
idna==3.10
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
multidict==6.4.3
mysql-connector-python==9.2.0
numpy @ file:///D:/bld/numpy_1745118932824/work/dist/numpy-2.2.5-cp313-cp313-win_amd64.whl#sha256=70b26fb49b3709b258c01f416229ce76fcfc2d669138e1e1cf7145e874630cf6
pg8000==1.31.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pymongo==4.11.3
pystac @ file:///home/conda/feedstock_root/build_artifacts/pystac_1744750342195/work
pystac-client==0.8.6
python-dateutil @ file:///C:/Users/dev-admin/perseverance-python-buildout/croot/python-dateutil_1729038406576/work
referencing==0.36.2
requests==2.32.3
rpds-py==0.24.0
rsa==4.9
scramp==1.4.5
shapely @ file:///D:/bld/shapely_1743678328769/work
six @ file:///C:/b/abs_149wuyuo1o/croot/six_1744271521515/work
SQLAlchemy==2.0.38
typing_extensions==4.13.2
urllib3==2.3.0
yarl==1.20.0
